requirements.txt
# janus_fa

absl-py==2.3.0
accelerate==1.7.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
attrdict==2.0.1
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
Brotli @ file:///croot/brotli-split_1736182456865/work
certifi @ file:///croot/certifi_1745939216646/work/certifi
cffi==1.17.1
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.2.1
colorama==0.4.5
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
datasets==2.14.4
debugpy==1.8.14
decorator==5.2.1
deepspeed==0.17.1
defusedxml==0.7.1
dill==0.3.7
docstring_parser==0.16
easydict==1.13
einops==0.8.1
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.115.12
fastjsonschema==2.21.1
ffmpy==0.6.0
filelock @ file:///croot/filelock_1744281381737/work
flash_attn==2.7.4.post1
fonttools==4.58.4
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2025.5.1
gmpy2 @ file:///croot/gmpy2_1738085463648/work
gradio==3.48.0
gradio_client==0.6.1
grpcio==1.73.0
h11==0.16.0
hf-xet==1.1.4
hjson==3.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
hydra-core==1.3.2
idna @ file:///croot/idna_1714398848350/work
importlib_resources==6.5.2
inquirerpy==0.3.4
ipykernel==6.29.5
ipython==8.37.0
ipywidgets==8.1.7
isoduration==20.11.0
jedi==0.19.2
Jinja2 @ file:///croot/jinja2_1741710844255/work
joblib==1.5.2
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
latex2mathml==3.78.0
lightning-utilities==0.14.3
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.5
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdtex2html==1.3.0
mdurl==0.1.2
mistune==3.1.3
mkl-service==2.4.0
mkl_fft @ file:///io/mkl313/mkl_fft_1730824109137/work
mkl_random @ file:///io/mkl313/mkl_random_1730823916628/work
mpmath @ file:///croot/mpmath_1690848262763/work
msgpack==1.1.1
multidict==6.4.4
multiprocess==0.70.15
narwhals==1.43.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx @ file:///croot/networkx_1737039604450/work
ninja==1.11.1.4
notebook==7.4.3
notebook_shim==0.2.4
numpy==1.26.4
nvidia-ml-py==12.575.51
omegaconf==2.3.0
open_clip_torch==2.32.0
orjson==3.10.18
overrides==7.7.0
packaging==25.0
pandas==2.3.0
pandocfilters==1.5.1
parso==0.8.4
peft==0.7.1
pexpect==4.9.0
pfzy==0.3.4
pillow==10.4.0
platformdirs==4.3.8
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.12.0
pyparsing==3.2.3
pypinyin==0.50.0
pyrootutils==1.0.4
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
python-multipart==0.0.20
pytorch-lightning==1.9.4
pytz==2025.2
PyYAML @ file:///croot/pyyaml_1728657952215/work
pyzmq==27.0.0
referencing==0.36.2
regex==2024.11.6
requests @ file:///croot/requests_1730999120400/work
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.2.0
rpds-py==0.25.1
safetensors==0.5.3
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==1.8.3
sentencepiece==0.1.96
shtab==1.7.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
starlette==0.46.2
sympy @ file:///croot/sympy_1738108488918/work
tensorboard==2.19.0
tensorboard-data-server==0.7.2
terminado==0.18.1
threadpoolctl==3.6.0
tiktoken==0.5.2
timm==1.0.14
tinycss2==1.4.0
tokenizers==0.15.2
tomli==2.2.1
torch==2.2.1
torchaudio==2.2.1
torchmetrics==1.7.3
torchvision==0.17.1
tornado==6.5.1
tqdm==4.64.0
traitlets==5.14.3
transformers==4.38.2
triton==2.2.0
trl==0.9.6
typeguard==4.4.3
types-python-dateutil==2.9.0.20250516
typing-inspection==0.4.1
typing_extensions==4.14.0
tyro==0.9.24
tzdata==2025.2
uri-template==1.3.0
urllib3 @ file:///croot/urllib3_1737133630106/work
uvicorn==0.34.3
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==3.1.3
widgetsnbextension==4.0.14
xxhash==3.5.0
yarl==1.20.1
